feat: move function stream util
